fix(csharp): resolve CodeQL cs/loss-of-precision + unused-container alerts
After the path-injection / integer-conversion fixes in 4ac94494,
CodeQL's next pass surfaced 7 remaining failure-level alerts in C#:

cs/loss-of-precision (6 alerts in ShortTextProcessor.cs:568 + 6 mirrors
in ShortTextProcessorTests.cs):
- (int)(sampleRate * SilencePadMs / 1000.0f) computed int*int before
  the float divide. CodeQL warned that the multiply could overflow
  before the cast made it float. Realistic sample rates make overflow
  impossible, but the warning is legitimate for theoretical inputs.
- Fix: (int)((long)sampleRate * SilencePadMs / 1000L). One operand
  promoted to long up front; pure integer math; no precision loss for
  any reachable sample rate.

cs/useless-assignment (1 alert in SwedishPhonemizerTests.cs:1678):
- Trigraph_Ckj_ProducesTjSound created `var input = new List<string> {
  "a" };` but never used it — the actual test calls
  _engine.ToPhonemeList("rackjobb"). Dead code from an earlier
  iteration. Removed along with the comment scaffolding that no longer
  applies.

The previous Python path-injection (10) and Go integer-conversion (1)
fixes in 4ac94494 stay in place; this batch addresses the pre-existing
C# debt that became visible once those moved off the top of the list.

+++ src/csharp/PiperPlus.Core.Tests/ShortTextProcessorTests.cs
@@ -878,7 +878,7 @@ public class ShortTextProcessorTests
     {
         short[] audio = [100, 200, 300];
         int sampleRate = 22050;
-        int expectedSilenceSamples = (int)(sampleRate * ShortTextProcessor.SilencePadMs / 1000.0f);
+        int expectedSilenceSamples = (int)((long)sampleRate * ShortTextProcessor.SilencePadMs / 1000L);
 
         short[] padded = ShortTextProcessor.PadSilenceForShortText(audio, sampleRate);
 
@@ -890,7 +890,7 @@ public class ShortTextProcessorTests
     {
         short[] audio = [100, 200, 300];
         int sampleRate = 22050;
-        int silenceSamples = (int)(sampleRate * ShortTextProcessor.SilencePadMs / 1000.0f);
+        int silenceSamples = (int)((long)sampleRate * ShortTextProcessor.SilencePadMs / 1000L);
 
         short[] padded = ShortTextProcessor.PadSilenceForShortText(audio, sampleRate);
 
@@ -906,7 +906,7 @@ public class ShortTextProcessorTests
     {
         short[] audio = [100, 200, 300];
         int sampleRate = 22050;
-        int silenceSamples = (int)(sampleRate * ShortTextProcessor.SilencePadMs / 1000.0f);
+        int silenceSamples = (int)((long)sampleRate * ShortTextProcessor.SilencePadMs / 1000L);
 
         short[] padded = ShortTextProcessor.PadSilenceForShortText(audio, sampleRate);
 
@@ -922,7 +922,7 @@ public class ShortTextProcessorTests
     {
         short[] audio = [100, 200, 300, 400, 500];
         int sampleRate = 22050;
-        int silenceSamples = (int)(sampleRate * ShortTextProcessor.SilencePadMs / 1000.0f);
+        int silenceSamples = (int)((long)sampleRate * ShortTextProcessor.SilencePadMs / 1000L);
 
         short[] padded = ShortTextProcessor.PadSilenceForShortText(audio, sampleRate);
 
@@ -952,7 +952,7 @@ public class ShortTextProcessorTests
         short[] padded = ShortTextProcessor.PadSilenceForShortText(audio, sampleRate);
 
         // Each silence block = sampleRate * SilencePadMs / 1000
-        int expectedSilenceSamples = (int)(sampleRate * ShortTextProcessor.SilencePadMs / 1000.0f);
+        int expectedSilenceSamples = (int)((long)sampleRate * ShortTextProcessor.SilencePadMs / 1000L);
         int expectedTotal = expectedSilenceSamples + 1 + expectedSilenceSamples;
         Assert.Equal(expectedTotal, padded.Length);
     }
@@ -965,7 +965,7 @@ public class ShortTextProcessorTests
 
         short[] padded = ShortTextProcessor.PadSilenceForShortText(audio, sampleRate);
 
-        int expectedSilenceSamples = (int)(sampleRate * ShortTextProcessor.SilencePadMs / 1000.0f);
+        int expectedSilenceSamples = (int)((long)sampleRate * ShortTextProcessor.SilencePadMs / 1000L);
         Assert.Equal(1 + (2 * expectedSilenceSamples), padded.Length);
     }
